unify components
